fix(desktop): move runtime detector below services
Config selects the desktop variant and basemap before service initialization, so importing the detector from services reversed the project dependency direction. Keep the canonical dependency-free detector in config and preserve the service path as a compatibility re-export for existing consumers and test mocks.
