+++ tests/runtime-env-guards.test.mjs
@@ -6,10 +6,10 @@ import { fileURLToPath } from 'node:url';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const runtimeSrc = readFileSync(resolve(__dirname, '../src/services/runtime.ts'), 'utf-8');
-// #5911 split the desktop detector (and its VITE_DESKTOP_RUNTIME read) into a
-// dependency-free leaf. The allow-list invariant follows the code: both halves
-// are asserted, so the split cannot be used to smuggle in a whole-env snapshot.
-const desktopRuntimeSrc = readFileSync(resolve(__dirname, '../src/services/desktop-runtime.ts'), 'utf-8');
+// The desktop detector (and its VITE_DESKTOP_RUNTIME read) lives in a
+// dependency-free config leaf. The allow-list invariant follows the canonical
+// implementation, so moving it cannot be used to smuggle in a whole-env snapshot.
+const desktopRuntimeSrc = readFileSync(resolve(__dirname, '../src/config/desktop-runtime.ts'), 'utf-8');
 const variantSrc = readFileSync(resolve(__dirname, '../src/config/variant.ts'), 'utf-8');
 
 describe('runtime env guards', () => {
